Logger hackathon - Minor updates on loggers under .runnable and namespacing of JSON exporter (#922)
* Updated logging
* Namespacing of JSON Exporter
